feat: 新番时间表及其圆屏适配 feat: 历史记录非视频显示类型，有番剧的自动跳转番剧 fix: 修复历史记录、首页推荐番剧会显示一个空的Up名字 fix: 搜索时没有点击反馈

+++ app/src/main/java/cn/spacexc/wearbili/dataclass/bangumi/EpisodeX.kt
@@ -1,7 +1,7 @@
 package cn.spacexc.wearbili.dataclass.bangumi
 
 data class EpisodeX(
-    val aid: Int,
+    val aid: Long,
     val badge: String,
     val badge_info: BadgeInfo,
     val badge_type: Int,
@@ -11,7 +11,7 @@ data class EpisodeX(
     val dimension: Dimension,
     val duration: Int,
     val from: String,
-    val id: Int,
+    val id: Long,
     val is_view_hide: Boolean,
     val link: String,
     val long_title: String,
