+++ app/src/main/java/cn/spacexc/wearbili/dataclass/bangumi/Series.kt
@@ -2,6 +2,6 @@ package cn.spacexc.wearbili.dataclass.bangumi
 
 data class Series(
     val display_type: Int,
-    val series_id: Int,
+    val series_id: Long,
     val series_title: String
 )
